# Remove executor dependency