@JosefWN 2022-07-13
Speaking of dependencies, we could also consider changing `hex` to `convert` (Dart's official library for conversions to hex and other formats), it's a bit bigger but it looks like it's better maintained and seems like a more optimized/less naive implementation at first glance?